Improve routing and add a dedicated 404 page
Routing is now more reactive and updates the view on any
route change, even invalid ones. This allows for a more
consistent user experience. A dedicated 404 page has been
added to handle cases where the user navigates to a non-existent route.
Also, the the routing logic now contains privileged routes
that are only accessible to users with admin status.