Backend: Internal Document API and Commits with Multiple Parents (#175)
This PR adds new functions to the internal Version Control API to work
with documents:
- `getDocument` gets a document by id
- `createDocument` creates a new document for a user group
- `createDocumentCommit` creates a new commit and tries to set it as the
new document head commit. This can fail if the new commit is unrelated
to the currenty set head commit of the document. This might change in
the future, as there are other cases in which it might be sensible to
fail the operation.

Also, commits can now have multiple parent commits. (yeah, this probably
should have been a separate PR)

---------

Co-authored-by: Finn Evers <dev@bahn.sh>