Load correct project ID in editor
When clicking a project in the project list of the home page,
the editor is now aware of the correct project id. Pressing the
"GET" button now fetches the correct project data (head commit
of the project). Because some projects do not have a head commit
yet, another debug button "forceGET" was added to force-load some
data from the server.