Disable scrollbar for completele pages
This commit removes the scrollbar from the main content area (all pages)
of the application. The user can still scroll, if necessary, but this is
necessary to improve the visual appearance of the application
(otherwise, the the scrollbar would push the navbar to the left on some
pages). The scrollbar is still visible in modals and other scrollable
containers (like the TOC or editor).