backend/ltml: Move flags up to proper nodes
Previously, flags could be set in leafs in the `Tree`, which would
result in flags in the AST wherever it was derived from a leaf of the
`Tree`.

Now, flags can be set at any node in the `Tree`, which are, again,
directly transferred to the corresponding locations in the AST.

In particular, headings (which are just annotated on nodes) can no
longer be flagged.  The same holds for the container header, where flags
never made sense.  Recall that flags are to indicate which subtree(s)
should be translated to HTML/PDF.

This change was desired s.t. the output (HTML, PDF) generators do not
need to generate output for headers.

The idea is that, now, when the frontend request an update w.r.t. a
heading (which it should request for the whole node), it gets back the
output for the whole subtree.

This change actually makes the code a little simpler / prettier, IMHO.

This also inspired extending the `SequenceF` typeclass to a
`TraversableF` typeclass, a proper generalization of `Traversable`.