Docs API: Create Document Now With Default Content (#482)
With this PR, new documents are initialized with default content.

However, creating a new document currently results in a frontend error.
This is most likely due to a change of the reponse data type of the api:
<img width="786" height="564" alt="grafik"
src="https://github.com/user-attachments/assets/be617bf7-86d7-4242-92c8-5c4123147646"
/>

The document is created successfully nevertheless. Since the generation
of a ToC is currently not possible, the headings of the default document
are not yet sensible in any way:
<img width="601" height="783" alt="grafik"
src="https://github.com/user-attachments/assets/45903d90-34eb-4f84-a11e-ea8d31d2a09c"
/>

Co-authored-by: Φnn Evers <dev@bahn.sh>