backend/ltml: Add static headings to nav toc
The navigation TOC now gets a header for every node.  It might be nicer
to hide some nodes (treat them as pseudo nodes)---specifically, the main
document body node if it is not a leaf---however, the chosen solution is
simpler.

The code currently does not work; it relies on a not-yet-existing
`mkRenderedTocEntry` function.