backend/ltml: Move static nav headings into AST
This way, the HTML code can directly generated all headings for the
navigation TOC.

Also, relatedly, added DocumentMainBody, DocumentIntro, and
DocumentExtro, plus corresponding types.

The static headings are no longer defined per kind, but per type, which
mostly means that they can be localized (not English).

The treeToMeta code is simplified, the AST and type AST are extended.
and the HTML heading generation code is to be changed.