backend/ltml: Move Parsed wrapper above Section
We previously artificially pushed the Parsed wrapper down into the
Heading and SectionBody, where it semantically does not belong, in order
to simplify handling by the output generators.

We have now decided to burden the output generators instead.

Note that the Heading remains wrapped in Parsed--for whenever we parse
the heading individually (when it is a non-leaf section).