backend/ltml: Add NavTocHeaded wrapper
Previosly, static nav toc headings were encoded as part of *Format; this
generally meant that in case of parse failure, the heading would be
lost--which is undesired, it is static after all.

Now, we use `NavTocHeaded` as a wrapper around `Parsed a` s.t. the
heading survives a parse failure.

This also prompted changing DocumentMainBody, DocumentIntro, and
DocumentExtro from newtype's to type aliases.