Improve ConsumableStackT monad transformer
Previously, it relied on a wrapped MonadFail instance, and delivered
static error messages.  Now, we return values of a custom error type.
This allows to define custom error messages, and seems more in line with
the common monad transformers.

It also does not wrap StateT anymore, but has a custom definition that
better matches its capabilities.

This is in anticipation for further use of ConsumableStackT.