backend/ltml: Wrap doc intro, extro in Maybe
This allows for easily omitting such nodes from the tree.

In particular, the UI now definitely won't show an editable intro/extro
that only accepts the empty input (modulo whitespace).

Note that it is still possible to permit such an empty intro/extro in
the LSD, but this should rarely be useful.