backend/ltml: Permit empty enum items
They may now also contain an initial sub-enumeration, which may be the
only content.

This is useful for intermediate editing.  No finished document should
have empty items or items containing only a sub-enumeration.  The latter
is why this was originally prohibited.  This has, however, been deemed
inconvenient.

Also
 - The (possibly labeled) keyword for an enumeration item may now be
   directly followed by text, where previously either at least one space
   or a newline was required.

The solution is not very pretty (in particular, the use of the
MiElementConfig record type with one field ignored), but works.
Ideally, during further MiTree / TextTree parser improvements, the
here-introduced changes should be improved.