backend/ltml: Permit some special chars in text
Newly allowed: `<`, `{`, `}`
Still unallowed: `>`, `\` (unless followed by valid escapable char)

This required some limited use of try; this only adds linear runtime,
and rarely much.  For styling (`<X...>`), only the `<X` is "tried"
(which is cheap), while for `{...}`, the whole element is "tried", but
this element is typically short, and never nested.

Note that `}` could have been safely allowed before, without any other
changes.

If we used try for the whole `<X...>`, we could also allow `>`, but that
is expensive.  It shall be permitted in another way.

The goal was to reduce parse errors.  See also #126.