Add meta-aware section insertion to TOC component
Now, sections are added in accordance with the document's meta map,
ensuring only valid section types can be inserted based on the current
context. This works for leaf sections.

This doesn't work for non-leaf sections yet. For example, when trying
to add a new "Einfaches Textdokument" to the appendix ("Anlagen"), the
structure of the created simple-document is not created. Only the top-
level section ("document", "simple-doc") is created, but the children
it needs according to the meta map. This is still a TODO!