Backend ltml universal tables (#722)
Implemented the logic to create tables inside LTML.

The syntax for such a table is defined as follows:

A cell is created with the keyword `|` and a row is terminated with the
keyword `|&`. In practice this may look like this:

```
| content00 | content01 | content02 |&
| content10 | content11 | content12 |&
```

It is also allowed to define a table like this: 

```
| content00 
| content01 
| content02 |&
| content10 
| content11 
| content12 |&
```

(The spaces before and after the content are optional.)

Further we want to allow merged cells, both horizontally, vertically and
blockwise.
To merge horizontally one writes `<` in all cells to the right of the
one with the content:
```
| content00 | <         |   <       |&
| content00 | content01 | content02 |&
```
which results in somthing like this:
```
|             content00             |
| content00 | content01 | content02 |
```
To merge vertically one writes `^` in all cells underneath the one with
the content:
```
| content00 | content01 | content02 |&
| ^         |   ^       | content12 |&
| ^         |   ^       | content22 |&
```
which in turn results in somthing like this:
```
|           |           | content02 |
| content00 | content01 | content12 |
|           |           | content22 |
```
To merge blockwise one writes the content in the topleft cell of the
box, all cells left of that must be `<` and all cells underneath must be
`^`. All other cells in the span can be either `<` or `^`:
```
| content00 | <         |   <       |&
| ^         |   ^       |   <       |&
| ^         |   <       |   ^       |&
```
which then results in:
```
|                                  |
|             content00            |
|                                  |
```
We only allow rectangular blocks, meaning somthing like:
```
| content00 | <         |   <       |&
| ^         |   ^       |   <       |&
| ^         |   <       | content22 |&
```
would result in:
```
|        content00      |   <       |
|                       |   <       |
| ^         |   <       | content22 |
```
where the merge symbols are simply parsed as text.

Further it is possible to insert highlighted cells by introducing them
with |* instead of |.

All of this is made compatible with the LSD and thus can be changed and
extended upon freely.
e.g.: it should be easily possible to introduce cells with typography,
like fontsize, alignment or a different background color. For now we
stick with default cells and highlighted ones, which have gray
background and a larger fontsize.

Closes #569

---------

Co-authored-by: Fabian Krull <60816643+Fab1000@users.noreply.github.com>