backend/latex: the story of automatically generating latex code (#757)
Since LaTeX is not as flexible as HTML when it comes to spacing, some
magic is needed to still be able to space everything properly in the
PDF.

One major problem is the spacing of the columns, which overflow quite
easily in LaTeX.
Since i found no way to integrate this properly in the automation and
there are still too many bugs with the multirows and multicolumns, i
decided to delegate some of the responsibility of making the table
"nice" to the user.

To that end a new optional syntax is introduced, where the user can
specify the proportions of the content. This can be done using the
following syntax at the top of the table:
```
|= 2 |= 5 |= 1 |= 1 |= 3
```
hereby the proportions are calculated relative to the total of the given
numbers (where columns without manual specification are simply give a
weight of 1. This allows the user to somewhat decide on the granularity
(simply by using larger numbers, even though this is to be taken with a
grain of salt.)

Other issues with the spacing, like overflowing content of multirows, or
cell content that still breaks the spacing of the table, must be handled
by inserting linebreaks manually.

I see that this may still not be a really neat solution, but I guess
LaTeX just isn't really made for this kind of "optimal" automization.
Another idea would be to render the table in HTML, export the result as
an image and insert this into the latex document. But of course that
would still cause some major inconveniences. E.g. how to handle longer
tables which would need a pagebreak? The only better solution would be
to base the tables even more on LaTeX syntax, giving the user more
flexibility. But of course this would stand in contrast to the intended
simplicity of the editor.